requirements.txt
pyrogram==2.0.106
tgcrypto
py-tgcalls==0.9.7
httpx
python-dotenv
